Reformatting and refactoring part 1
