+++ frontend/udlånssystem/src/components/Nav-btn.svelte
@@ -12,6 +12,7 @@
 
   function handleClick() {
     navigate(destination, { replace: true });
+    
     path.update(() => {
       return destination;
     });
@@ -20,7 +21,6 @@
 
 <button
   on:click={handleClick}
-  
   class:selected={stripPath(currentPath, 1) === destination}
 >
   <div class="flex">
@@ -28,7 +28,6 @@
     <p>{text}</p>
   </div>
 </button>
-
 
 <style>
   button {
@@ -43,8 +42,9 @@
   }
   button:hover:not(.selected) {
     background-color: var(--bg2);
-    
-    box-shadow: rgba(0, 0, 0, 0.25) 0px 0.0625em 0.0625em,
+
+    box-shadow:
+      rgba(0, 0, 0, 0.25) 0px 0.0625em 0.0625em,
       rgba(0, 0, 0, 0.25) 0px 0.125em 0.5em,
       rgba(255, 255, 255, 0.1) 0px 0px 0px 1px inset;
   }
